requirements.txt
streamlit
requests
numpy
pandas
matplotlib
time
random
geopy.geocoders
